Enhance README with Scripting Task details
Create a shell script where the script should analyze the health of the virtual machine based on cpu, memory and disk space. If any of these are less than 60% utilized , the script declares the state of vm as healthy whereas if any of these parameters are more than 60% , script declares the health as not healthy.
Also, the script should support a command line argument named "explain", when passed, the script should explain the reason for health status along with printing the health status. Target vm's are always ubuntu.